MultiVAC/requirements.txt
Flask==0.10.1
Flask-Bootstrap==3.3.5.6
Flask-DebugToolbar==0.10.0
flask-mongoengine==0.7.1
flask-paginate==0.4.1
Flask-RQ==0.2
Flask-Runner==2.1.1
Flask-Script==2.0.5
Flask-Testing==0.4.2
Flask-WTF==0.12
Flask-Mail==0.9.0
Jinja2==2.8
pymongo==3.2.2
python-dateutil==2.4.2
python-dotenv==0.3.0
python-magic==0.4.10
redis==2.10.5
requests[security]==2.9.1
